moved JMH states from `DBBenchmark` to dedicated files
