+++ benchmark/src/main/scala/org/alephium/explorer/benchmark/db/state/ReadBenchmarkState.scala
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with the library. If not, see <http://www.gnu.org/licenses/>.
 
-package org.alephium.explorer.benchmark.db
+package org.alephium.explorer.benchmark.db.state
 
 import scala.annotation.tailrec
 import scala.reflect.ClassTag
@@ -23,6 +23,8 @@ import com.typesafe.scalalogging.StrictLogging
 import org.openjdk.jmh.annotations._
 import slick.basic.DatabaseConfig
 import slick.jdbc.JdbcProfile
+
+import org.alephium.explorer.benchmark.db.DBExecutor
 
 /**
   * Base implementation for JMH states, for benchmarking reads queries to the target database.
