+++ benchmark/src/test/scala/org/alephium/explorer/benchmark/db/state/WriteBenchmarkStateSpec.scala
@@ -14,11 +14,12 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with the library. If not, see <http://www.gnu.org/licenses/>.
 
-package org.alephium.explorer.benchmark.db
+package org.alephium.explorer.benchmark.db.state
 
 import scala.concurrent.duration._
 
 import org.alephium.explorer.AlephiumSpec
+import org.alephium.explorer.benchmark.db.DBExecutor
 import org.alephium.explorer.benchmark.db.util.TestUtils._
 
 /**
